Added Blockage option in protobuf netlist and RePlAce Flow
